% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bg.r
\name{bg}
\alias{bg}
\title{Build a BioGRID query}
\usage{
bg(access_point)
}
\arguments{
\item{access_point}{what to query for (character string). See
details for available options.}

\item{key}{if you wish to explicitly set an access key (character string). If
left as NULL will use the key set by \code{\link{bg_get_key}}.}
}
\description{
Build a BioGRID query
}

