% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contrain.R
\name{bg_constrain}
\alias{bg_constrain}
\title{contrain BioGRID query which additional parameters}
\usage{
bg_constrain(bg, ...)
}
\arguments{
\item{bg}{bg object}

\item{...}{additional parameters}
}
\value{
modified bg object. N.B. current behaviour overwites and previously
present params.
}
\description{
contrain BioGRID query which additional parameters
}

