% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contrain.R
\name{bg_constrain_params}
\alias{bg_constrain_params}
\title{show data.frame of available constrain parameters}
\usage{
bg_constrain_params()
}
\description{
show data.frame of available constrain parameters
}
\keyword{internal}

