% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_key.R
\name{bg_get_key}
\alias{bg_get_key}
\title{GET an access key for BioGRID web service}
\usage{
bg_get_key(first, last, email, project)
}
\arguments{
\item{first}{first name (character string).}

\item{last}{last name (character string).}

\item{email}{e-mail (character string).}

\item{project}{project name (character string).}
}
\value{
(character string) Biogrid access key invisibly.
}
\description{
GET an access key for BioGRID web service
}
\details{
To submit requests to the BioGRID web service an application key is required.
You can obtain a key from \url{http://webservice.thebiogrid.org/} or via this
function. According to BioGRID this info is only used for notifying you in
the event of a problem.
}
\section{I already have a key}{

If you have already done this it just returns the key associated with those
credentials.
}

