% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_results.R
\name{bg_get_results}
\alias{bg_get_results}
\title{GET results of a query to the BioGRID web service.}
\usage{
bg_get_results(bg, as = "tab2", file = NULL, .request = FALSE)
}
\arguments{
\item{bg}{biogrid object returned by \code{\link{bg}}. (named list)}

\item{as}{format in which to return results. (character string)}

\item{file}{Specify name of file if you wish to write the results to file.
(character string)}
}
\description{
GET results of a query to the BioGRID web service.
}

