% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infer.R
\name{get_organisms}
\alias{get_organisms}
\title{Organisms supported by BioGrid}
\usage{
get_organisms()
}
\value{
data.frame of organisms and their taxID supported by biogrid
}
\description{
Organisms supported by BioGrid
}
\keyword{internal}

