% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_key.R
\name{random_string}
\alias{random_string}
\title{generate a random character string}
\usage{
random_string(length = 10)
}
\arguments{
\item{length}{length of character string}
}
\description{
generate a random character string
}
\keyword{internal}

