% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_results.R
\name{ratify_bg}
\alias{ratify_bg}
\title{Validate the structure bg object}
\usage{
ratify_bg(bg)
}
\arguments{
\item{bg}{bg}
}
\value{
bg object (named list)
}
\description{
Validate the structure bg object
}
\keyword{internal}

