% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_results.R
\name{ratify_format}
\alias{ratify_format}
\title{Validate if a format is allowed}
\usage{
ratify_format(format)
}
\arguments{
\item{format}{format}
}
\value{
format type (character string)
}
\description{
Validate if a format is allowed
}
\keyword{internal}

