% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bg.r
\name{ratify_key}
\alias{ratify_key}
\title{validate application key and retrieve it}
\usage{
ratify_key()
}
\value{
application key (character string)
}
\description{
validate application key and retrieve it
}
\keyword{internal}

