% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_results.R
\name{ratify_request}
\alias{ratify_request}
\title{Validate GET request}
\usage{
ratify_request(request)
}
\arguments{
\item{request}{request}
}
\value{
TRUE if successful or list of status code and content.
}
\description{
Validate GET request
}
\keyword{internal}

