% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse.R
\name{select_parser_fun}
\alias{select_parser_fun}
\title{specify how to parse request contents}
\usage{
select_parser_fun(bg)
}
\arguments{
\item{bg}{bg}
}
\description{
specify how to parse request contents
}
\keyword{internal}

