% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/contrain.R
\name{stratify_parameters}
\alias{stratify_parameters}
\title{validate constrain parameters}
\usage{
stratify_parameters(param_list)
}
\arguments{
\item{param_list}{list of parameters}
}
\value{
parameter list if valid, stop if not.
}
\description{
validate constrain parameters
}
\keyword{internal}

