% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/infer.R
\name{taxId}
\alias{taxId}
\title{retreive taxonomy id for through fuzzy matching}
\usage{
taxId(organism)
}
\arguments{
\item{organism}{name (character string)}
}
\value{
taxonomy id (integer vector)
}
\description{
retreive taxonomy id for through fuzzy matching
}
\examples{
taxId("Drosophila melanogaster")
}

